apps/ai_run_api/requirements.txt
# --- Web API (FastAPI) ---
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
python-multipart>=0.0.9
pydantic>=2.6,<3.0

# --- Env / Config ---
python-dotenv>=1.0.1

# --- HTTP client (파일 다운로드 / Clova OCR) ---
httpx>=0.27.0

# --- File / Data parsing ---
pandas>=2.1.4
openpyxl>=3.1.2
PyMuPDF>=1.24.0

# --- LLM (OpenAI) ---
openai>=1.30.0

# --- Test ---
pytest>=7.4.0
pytest-asyncio>=0.23.0
